requirements.txt
dill==0.3.6
numpy==1.24.3
pandarallel==1.6.5
pandas==2.0.2
Pillow==9.5.0
pydantic==1.10.8
scipy==1.10.1
tqdm==4.65.0
